Rename hasher to matcher and add utiltest
