+++ python-threatexchange/threatexchange/extensions/video_vpdq/tests/test_vpdq_hash_and_match.py
@@ -1,4 +1,5 @@
 # Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
+
 import pytest
 from pathlib import Path
 
@@ -14,8 +15,10 @@ else:
         vpdq_to_json,
         json_to_vpdq,
         read_file_to_hash,
+        dedupe,
+        quality_filter,
     )
-    from threatexchange.extensions.video_vpdq.vpdq_brute_hasher import (
+    from threatexchange.extensions.video_vpdq.vpdq_brute_matcher import (
         match_VPDQ_hash_brute,
     )
 
@@ -26,6 +29,16 @@ ROOTDIR = Path(__file__).parents[5]
 
 @pytest.mark.skipif(_DISABLED, reason="vpdq not installed")
 class TestVPDQHasherMatcher:
+    def test_vpdq_utils(self):
+        example_hash = VideoVPDQSignal.get_examples()[0]
+        example_features = json_to_vpdq(example_hash)
+        feature_size = len(example_features)
+        assert len(dedupe(example_features)) == feature_size
+        assert len(quality_filter(example_features, 50)) == feature_size
+        assert len(quality_filter(example_features, 101)) == 0
+        example_features.append(example_features[0])
+        assert len(dedupe(example_features)) == feature_size
+
     def test_vpdq_from_string_path(self):
         computed_hash = VideoVPDQSignal.hash_from_file(ROOTDIR / VIDEO)
         expected_hash = read_file_to_hash(ROOTDIR / HASH)
