+++ python-threatexchange/threatexchange/extensions/video_vpdq/vpdq_util.py
@@ -1,4 +1,5 @@
 # Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
+
 import vpdq
 import json
 from json import JSONEncoder
